Move scrollerContents styles into HorizontalScroller/styles
